grow_to methods accepts `zero_memory` argument
The users of the PinnedVec trait, hence SplitVec, might require to zero out allocated memory which is not yet written. This is an additional safety guarantee for specifically concurrent data types. Therefore, grow_to and concurrently_grow_to methods accept a second argument zero_memory. When true is passed in, the implementor is responsible for zeroing out the new memory, if allocated.

* `Fragment::zero(&mut self)` method is implemented. This method is used on new fragments when grow to methods are called with a `zero_memory` of true.
* growth with memory zeroing tests are added.
* Tests extended with the `test-case` crate.